PR-06: extract pause.rs and focus_pipeline.rs
Behaviour-preserving relocation:
- pause.rs (new): UnpauseContext, local_sni_unpause_context,
  pause_daemon, unpause_daemon, plus test-only TEST_LAST_UNPAUSE_REQUEST_ENV
  and its helpers.
- focus_pipeline.rs (new): execute_focus_actions, extract_focus_layer,
  update_status_for_focus, handle_focus_event, native_terminal_window,
  resolve_sni_focus_only.

main.rs: 5791 -> 5579 LOC. Visibility widenings (D19 + extras):
query_focus_for_env, apply_focus_for_env, SniSettingsStore,
SNI_DEFAULT_SHOW_FOCUS_ONLY are now pub(crate) so the new modules
can reach still-in-main.rs consumers. apply/query_focus_for_env
revert when they move to backends/mod.rs in PR-11; SNI items
move in PR-14.

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
